requirements.txt
fastapi
uvicorn
python-telegram-bot
supabase
openai
